parsers: remove base() method
It was used to only access the name of the Parser. It returned a ParserBase,
but ParserBase has some fields that won't be needed to all Parser types.

To support more Parser implementations (especially EStringParser) we have
to remove the ParserBase contstraint. It will be splitted into two parts and
the base() and base_mut() methods will be removed.

Signed-off-by: Tibor Benke <ihrwein@gmail.com>