grammar: add GREEDY parser
The GREEDY parser tries to fill in the gap between two literals. It uses
the EStringParser and sets it's end_string parameter es the next literal
in the pattern.

The GREEDY parser doesn't fit well into the set of currently implemented
parsers. It needs to have some knowledge about it's environment (the next
literal) so it cannot be placed among the other parsers. It also returns
two TokenType instances, the parser and the following literal, so I had to
refactor the current grammar.  From now every parser can return arbitrary
number of TokenTypes, although all but the GREEDY parser returns only one.

Signed-off-by: Tibor Benke <ihrwein@gmail.com>