matcher: create Matcher struct with integration tests
The Matcher struct is a high level structure to create a parser  instance.
It's purpose its to abstract away the details of parser creation and to
provide a simple interface to the users. Currently it has two methods:

* from_file() creates a parser, where the patterns are listed in a file
* parse() can be used to parse texts into key-value pairs

It uses the try!() macro in several places, its documentation is worth a
check: https://doc.rust-lang.org/std/macro.try!.html

Signed-off-by: Tibor Benke <tibor.benke@balabit.com>