Matcher: implement Clone trait
It is needed by syslog-ng's logpipe clone() method. This way a parser can
be present in multiple logpaths.

Every other part of a Matcher needs to implement Clone, that's why the
other changes.

Note, that ParserNode manually implements the Clone trait and the Parser
trait is extended with a boxed_clone() method. The Parser trait cannot
implement Clone because it's not object safe and that's screw up
everything. The boxed_clone() method can be extracted into it's own trait
but it only makes sense to use it with parameters, like BoxedClone<Parser>.
That solution was rejected by the compiler.

Signed-off-by: Tibor Benke <ihrwein@gmail.com>