source: return BuildResult instead of Pattern
From now on the PatternSource iterator returns BuildResult instances
instead of Patterns. This makes possible to fail when an Pattern cannot
be parsed properly in an iteration. Before this modification we returned
a Some(Pattern) or None. From now we return a Result<Pattern, BuildError>
enum so the errors can be reported easily.

Signed-off-by: Tibor Benke <ihrwein@gmail.com>