matcher::Factory: add from_file() function
The Factory::from_file() function can be used to chose the right
implementation to open pattern files. When a file has a "json"
extension it will be opened by SerializedPatternFile. If it has
other extension it will be opened by PlainPatternFile.

Files without extensions are not supported and will cause error.

Signed-off-by: Tibor Benke <ihrwein@gmail.com>