TrieParserFactory: set optional parameters in TrieParserFactory
This little change decouples the grammar and the Parser instances. A min_len
keyword is part of the grammar and the parsers shouldn't know about
the keyword, just its value. This change moves the optional parameter
set logic into the TrieParserFactory which is the interface between the
grammar and the Parsers.

Signed-off-by: Tibor Benke <ihrwein@gmail.com>